feat(auth): implement Google login processing and remove obsolete GoogleAuth module
