+++ src/app.module.ts
@@ -14,10 +14,8 @@ import { BrandsModule } from './modules/brands/brands.module';
 import { ModelsModule } from './modules/models/models.module';
 import { VersionsModule } from './modules/versions/versions.module';
 import { PassportModule } from '@nestjs/passport';
-import { GoogleAuthModule } from './modules/googleAuth/google-auth.module';
 
 dotenv.config({ path: '.env.development' });
-
 
 @Module({
   imports: [
@@ -33,10 +31,10 @@ dotenv.config({ path: '.env.development' });
       },
     }),
     JwtModule.register({
-    global: true,
-    signOptions: { expiresIn: "1d" },
-    secret: process.env.JWT_SECRET,
-  }),
+      global: true,
+      signOptions: { expiresIn: '1d' },
+      secret: process.env.JWT_SECRET,
+    }),
     PassportModule.register({ defaultStrategy: 'jwt' }),
     AuthModule,
     UsersModule,
@@ -47,11 +45,8 @@ dotenv.config({ path: '.env.development' });
     ModelsModule,
     VersionsModule,
     YearOptionsModule,
-    GoogleAuthModule,
   ],
   controllers: [],
   providers: [],
 })
 export class AppModule {}
-
-
