--- src/dto/usersDto/complete-profile.dto.ts
@@ -1,24 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { IsOptional, IsString, IsNumber } from 'class-validator';
-
-export class CompleteProfileDto {
-  @ApiProperty({ required: false })
-  @IsOptional()
-  @IsNumber()
-  phone?: number;
-
-  @ApiProperty({ required: false })
-  @IsOptional()
-  @IsString()
-  country?: string;
-
-  @ApiProperty({ required: false })
-  @IsOptional()
-  @IsString()
-  city?: string;
-
-  @ApiProperty({ required: false })
-  @IsOptional()
-  @IsString()
-  address?: string;
-}
