+++ src/dto/usersDto/google-profile.dto.ts
@@ -1,0 +1,13 @@
+import { IsEmail, IsString } from 'class-validator';
+
+export class GoogleProfileDto {
+  @IsString()
+  sub: string;
+
+  @IsString()
+  name: string;
+
+  @IsString()
+  @IsEmail()
+  email: string;
+}
