--- src/modules/googleAuth/google-auth.module.ts
@@ -1,19 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PassportModule } from '@nestjs/passport';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from 'src/entities/user.entity';
-import { UsersModule } from '../users/users.module';
-import { GoogleAuthController } from './google-auth.controller';
-import { GoogleAuthService } from './google-auth.service';
-
-@Module({
-  imports: [
-    PassportModule.register({ defaultStrategy: 'jwt' }),
-    UsersModule,
-    TypeOrmModule.forFeature([User]),
-  ],
-  controllers: [GoogleAuthController],
-  providers: [GoogleAuthService],
-  exports: [GoogleAuthService],
-})
-export class GoogleAuthModule {}
